# contracts/api.yaml
...
openapi: 3.0.3
info:
  title: "AML Case Management API"
  version: "1.0.0"
  description: |
    API contracts for the AML Case Management System.

    ## Overview

    This API enables Spriggy to meet AUSTRAC AML/CTF compliance obligations by providing
    endpoints for managing KYC remediation, PEP/Sanctions screening disposition, and
    suspicious activity investigation workflows.

    ## Authentication

    All endpoints require Bearer token authentication via OIDC SSO integration (D5).
    Include the token in the Authorization header:

    ```
    Authorization: Bearer <access_token>
    ```

    Webhook endpoints use HMAC signature validation instead of Bearer tokens (D11).

    ## Roles

    The system enforces four distinct roles (FR-063):
    - **L1_ANALYST**: Triage and basic case handling
    - **L2_ANALYST**: Full investigation and SMR recommendation authority
    - **AML_MANAGER**: Approval authority for SMRs
    - **READ_ONLY**: View and export only

    ## Error Handling

    All errors follow a consistent format with machine-readable codes.
    See the ErrorResponse schema for details.

    ## Optimistic Locking

    Case mutations require a `version` field for optimistic locking (D8).
    Version mismatch returns 409 Conflict.

    ## Constitution Alignment

    - **Principle I**: All mutating endpoints create audit log entries
    - **Principle II**: Endpoints enforce role-based access
...
  contact:
    name: AML Compliance Team
...
servers:
  - url: /api/v1
...
paths:
...
  /cases:
    get:
      summary: List cases
      description: |
        Retrieve paginated list of cases with filtering.
        Default ordering: SLA urgency (most urgent first), then creation date (oldest first).

        **Roles**: All authenticated users

        **Source Requirements**: FR-024, FR-025, US-12
      operationId: listCases
      tags:
        - Cases
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          description: Filter by case status
          schema:
            $ref: '#/components/schemas/CaseStatus'
        - name: caseType
          in: query
          description: Filter by case type
          schema:
            $ref: '#/components/schemas/CaseType'
        - name: assignedUserId
          in: query
          description: Filter by assigned analyst (use 'me' for current user, 'unassigned' for queue)
          schema:
            type: string
        - name: tier
          in: query
          description: Filter by analyst tier
          schema:
            $ref: '#/components/schemas/Tier'
        - name: slaStatus
          in: query
          description: Filter by SLA status
          schema:
            type: string
            enum: [on_track, warning, breached]
        - name: customerId
          in: query
          description: Filter by customer ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of cases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /cases/{caseId}/investigation-findings:
    parameters:
      - $ref: '#/components/parameters/CaseIdParam'

    get:
      summary: Get investigation findings
      description: |
        Retrieve investigation findings for a case.

        **Roles**: All authenticated users

        **Source Requirements**: US-4
      operationId: getInvestigationFindings
      tags:
        - Investigation
      responses:
        '200':
          description: Investigation findings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestigationFindingsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Case or findings not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create investigation findings
      description: |
        Document investigation findings for a case.

        **Precondition**: Findings do not already exist for this case.

        **Roles**: L2_ANALYST, AML_MANAGER

        **Source Requirements**: US-4, BR-INV-001
      operationId: createInvestigationFindings
      tags:
        - Investigation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvestigationFindingsRequest'
      responses:
        '201':
          description: Investigation findings created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestigationFindingsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: L2+ role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Investigation findings already exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update investigation findings
      description: |
        Update existing investigation findings.

        **Roles**: L2_ANALYST, AML_MANAGER

        **Source Requirements**: US-4
      operationId: updateInvestigationFindings
      tags:
        - Investigation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInvestigationFindingsRequest'
      responses:
        '200':
          description: Investigation findings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestigationFindingsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # --------------------------------------------------------------------------
  # SMR Workflow
  # --------------------------------------------------------------------------
...
  /notifications:
    get:
      summary: List user notifications
      description: |
        Retrieve notifications for the current user.

        **Polling**: Designed for 30-second polling interval (D6).

        **Roles**: All authenticated users

        **Source Requirements**: FR-066, FR-067, FR-068, FR-069
      operationId: listNotifications
      tags:
        - Notifications
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: unreadOnly
          in: query
          description: Filter to unread notifications only
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: User notifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /notifications/{notificationId}/read:
    parameters:
      - name: notificationId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      summary: Mark notification as read
      description: |
        Mark a notification as read.

        **Roles**: All authenticated users
      operationId: markNotificationRead
      tags:
        - Notifications
      responses:
        '200':
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /notifications/read-all:
    post:
      summary: Mark all notifications as read
      description: |
        Mark all unread notifications as read.

        **Roles**: All authenticated users
      operationId: markAllNotificationsRead
      tags:
        - Notifications
      responses:
        '200':
          description: All notifications marked read
          content:
            application/json:
              schema:
                type: object
                properties:
                  markedReadCount:
                    type: integer
                    example: 12
        '401':
          $ref: '#/components/responses/Unauthorized'

  # --------------------------------------------------------------------------
  # Reports
  # --------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from OIDC SSO

  # --------------------------------------------------------------------------
  # Common Parameters
  # --------------------------------------------------------------------------
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CaseIdParam:
      name: caseId
      in: path
      required: true
      description: Case unique identifier
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: One or more fields are invalid
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Authentication required
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: You do not have permission to access this resource
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: The requested resource was not found
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable message
          example: One or more fields are invalid
        details:
          type: object
          additionalProperties: true
          description: Additional error context
...
    CaseType:
      type: string
      enum:
        - KYC_REMEDIATION
        - PEP_SANCTIONS_SCREENING
        - SUSPICIOUS_ACTIVITY
        - SUPPLEMENTARY_SMR
    CaseSubtype:
      type: string
      enum:
        - SANCTIONS_ONBOARDING
        - SANCTIONS_EXISTING_CUSTOMER
        - PEP_EDD_REQUIRED
        - PEP_PROVISIONAL_REVIEW
    CaseStatus:
      type: string
      enum:
        - OPEN
        - IN_PROGRESS
        - PENDING_INFORMATION
        - ESCALATED
        - PENDING_APPROVAL
        - CLOSED
        - SMR_FILED
    L2ReviewStatus:
      type: string
      enum:
        - NOT_REQUIRED
        - PENDING_REVIEW
        - REVIEWED_ACCEPTED
        - REVIEWED_REOPENED
    Tier:
      type: string
      enum:
        - L1
        - L2
    UserRole:
      type: string
      enum:
        - L1_ANALYST
        - L2_ANALYST
        - AML_MANAGER
        - READ_ONLY
...
    InvestigationConclusion:
      type: string
      enum:
        - NO_SUSPICION
        - SUSPICIOUS_ACTIVITY_IDENTIFIED
        - INSUFFICIENT_EVIDENCE
        - FURTHER_MONITORING_REQUIRED
...
    NotificationType:
      type: string
      enum:
        - CASE_ASSIGNED
        - SLA_WARNING
        - SLA_BREACH
        - ESCALATION
        - SMR_PENDING_APPROVAL
        - SMR_APPROVED
        - SMR_REJECTED
...
    CaseSummary:
      type: object
      required:
        - id
        - referenceCode
        - caseType
        - status
        - tier
        - slaDeadline
        - slaStatus
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        referenceCode:
          type: string
          example: AML-1001
        caseType:
          $ref: '#/components/schemas/CaseType'
        caseSubtype:
          $ref: '#/components/schemas/CaseSubtype'
        status:
          $ref: '#/components/schemas/CaseStatus'
        tier:
          $ref: '#/components/schemas/Tier'
        slaDeadline:
          type: string
          format: date-time
        slaStatus:
          type: string
          enum: [on_track, warning, breached]
        slaBreached:
          type: boolean
        l2ReviewStatus:
          $ref: '#/components/schemas/L2ReviewStatus'
        assignedUser:
          $ref: '#/components/schemas/UserSummary'
        customer:
          $ref: '#/components/schemas/CustomerSummary'
        createdAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
...
    CaseListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CaseSummary'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...
    CustomerSummary:
      type: object
      description: |
        Customer snapshot captured at case creation time.

        **Immutability**: Customer data is captured as a snapshot when the case is created.
        Subsequent changes to the customer's profile in Spriggy do not update this snapshot.
        This ensures case context reflects the state at the time of the alert, which is
        important for audit and investigation purposes. If updated customer information
        is needed, analysts can reference Spriggy's customer management system directly.
      required:
        - id
        - externalId
      properties:
        id:
          type: string
          format: uuid
        externalId:
          type: string
          description: Spriggy customer identifier
        displayName:
          type: string
          description: Masked or authorized display name
        accountStatus:
          type: string
          enum: [ACTIVE, SUSPENDED, CLOSED]
          description: Customer account status at time of case creation (snapshot)
        onboardingStatus:
          type: string
          enum: [IN_PROGRESS, COMPLETED, BLOCKED, UNKNOWN]
          description: Customer onboarding status at time of case creation (snapshot)
        snapshotAt:
          type: string
          format: date-time
          description: Timestamp when this customer data snapshot was captured

    # --- User Schemas ---
    UserSummary:
      type: object
      required:
        - id
        - displayName
        - role
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
...
    InvestigationFindingsResponse:
      type: object
      required:
        - id
        - caseId
        - conclusion
        - completedAt
        - completedBy
      properties:
        id:
          type: string
          format: uuid
        caseId:
          type: string
          format: uuid
        entitiesInvolved:
          type: array
          items:
            $ref: '#/components/schemas/InvolvedEntity'
        transactionPatterns:
          type: array
          items:
            $ref: '#/components/schemas/TransactionPattern'
        suspiciousIndicators:
          type: array
          items:
            $ref: '#/components/schemas/SuspiciousIndicator'
        conclusion:
          $ref: '#/components/schemas/InvestigationConclusion'
        notes:
          type: string
        completedAt:
          type: string
          format: date-time
        completedBy:
          $ref: '#/components/schemas/UserSummary'
    InvolvedEntity:
      type: object
      required:
        - entityType
        - name
      properties:
        entityType:
          type: string
          enum: [PERSON, BUSINESS, ACCOUNT, OTHER]
        identifier:
          type: string
        name:
          type: string
        role:
          type: string
        notes:
          type: string
    TransactionPattern:
      type: object
      required:
        - patternType
        - description
      properties:
        patternType:
          type: string
        description:
          type: string
        dateRange:
          type: object
          properties:
            from:
              type: string
              format: date
            to:
              type: string
              format: date
        totalAmount:
          type: number
        transactionCount:
          type: integer
    SuspiciousIndicator:
      type: object
      required:
        - indicatorCode
        - description
        - severity
      properties:
        indicatorCode:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        evidence:
          type: string
    CreateInvestigationFindingsRequest:
      type: object
      required:
        - conclusion
      properties:
        entitiesInvolved:
          type: array
          items:
            $ref: '#/components/schemas/InvolvedEntity'
        transactionPatterns:
          type: array
          items:
            $ref: '#/components/schemas/TransactionPattern'
        suspiciousIndicators:
          type: array
          items:
            $ref: '#/components/schemas/SuspiciousIndicator'
        conclusion:
          $ref: '#/components/schemas/InvestigationConclusion'
        notes:
          type: string
    UpdateInvestigationFindingsRequest:
      type: object
      properties:
        entitiesInvolved:
          type: array
          items:
            $ref: '#/components/schemas/InvolvedEntity'
        transactionPatterns:
          type: array
          items:
            $ref: '#/components/schemas/TransactionPattern'
        suspiciousIndicators:
          type: array
          items:
            $ref: '#/components/schemas/SuspiciousIndicator'
        conclusion:
          $ref: '#/components/schemas/InvestigationConclusion'
        notes:
          type: string

    # --- SMR Schemas ---
...
    NotificationResponse:
      type: object
      required:
        - id
        - notificationType
        - title
        - body
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        notificationType:
          $ref: '#/components/schemas/NotificationType'
        title:
          type: string
        body:
          type: string
        caseId:
          type: string
          format: uuid
        caseReference:
          type: string
        readAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    NotificationListResponse:
      type: object
      required:
        - data
        - pagination
        - unreadCount
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/NotificationResponse'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
        unreadCount:
          type: integer

    # --- Report Schemas ---
...
    PaginationMeta:
      type: object
      required:
        - page
        - limit
        - total
        - totalPages
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8
        hasNext:
          type: boolean
        hasPrev:
          type: boolean

# ============================================================================
# DEFAULT SECURITY
# ============================================================================
security:
  - bearerAuth: []